transit-routing/requirements.txt
# Core dependencies
numpy>=1.26.0
pandas>=2.1.0

# Scientific computing
scipy>=1.11.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# Graph algorithms
networkx>=3.2

# Data validation
pydantic>=2.0.0

# Configuration
pyyaml>=6.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code quality
black>=23.3.0
flake8>=6.0.0
mypy>=1.4.0

# Jupyter
jupyter>=1.0.0
notebook>=7.0.0
ipykernel>=6.25.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0

# Type hints
types-PyYAML>=6.0.12

# Database
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0  # PostgreSQL

# Flask
Flask>=3.0.0
Flask-CORS>=4.0.0

# swagger-api
flasgger

# for navigation + session
flask-socketio==5.3.4
python-socketio==5.9.0
eventlet==0.33.3
redis==5.0.1
celery==5.3.4